% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-stepribbon.r
\docType{data}
\name{geom_stepribbon}
\alias{geom_stepribbon}
\alias{GeomStepribbon}
\title{Step ribbon plots.}
\usage{
geom_stepribbon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  kmplot = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If 
specified and inherit.aes = TRUE (the default), it is combined with the
default mapping at the top level of the plot. You must supply mapping if
there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{na.rm}{If FALSE, the default, missing values are removed with a
warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
NA, the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them. This is most useful for helper functions that define
both data and aesthetics and shouldn't inherit behaviour from the default
plot specification, e.g. \code{borders()}.}

\item{kmplot}{If \code{TRUE}, missing values are replaced by the previous
values. This option is needed to make Kaplan-Meier plots if the last
observation has event, in which case the upper and lower values of the
last observation are missing. This processing is optimized for results
from the survfit function.}

\item{...}{Other arguments passed on to layer()'s params argument.}
}
\description{
\code{geom_stepribbon} is an extension of the \code{geom_ribbon}, and
is optimized for Kaplan-Meier plots with pointwise confidence intervals
or a confidence band.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "ribbon")}
}

\examples{
huron <- data.frame(year = 1875:1972, level = as.vector(LakeHuron))
h <- ggplot(huron, aes(year))
h + geom_stepribbon(aes(ymin = level - 1, ymax = level + 1), fill = "grey70") +
    geom_step(aes(y = level))
h + geom_ribbon(aes(ymin = level - 1, ymax = level + 1), fill = "grey70") +
    geom_line(aes(y = level))
}
\seealso{
\code{\link[ggplot2]{geom_ribbon}} \code{geom_stepribbon}
  inherits from \code{geom_ribbon}.
}
\keyword{datasets}
