\name{kmg2_box-class}
\docType{class}
\alias{kmg2_box}
\alias{kmg2_box-class}
\alias{plotBox,kmg2_box-method}
\alias{kmg2_plot_box}
\title{Box plot}
\description{
  This function calls an R Commander plug-in window for Box plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
