\name{kmg2_contour-class}
\docType{class}
\alias{kmg2_contour}
\alias{kmg2_contour-class}
\alias{plotContour,kmg2_contour-method}
\alias{kmg2_plot_contour}
\title{Contour plot}
\description{
  This function calls an R Commander plug-in window for contour plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}},
  \code{\link{kmg2_contour_df}}
}
\keyword{hplot}
