\name{kmg2_discretebar-class}
\docType{class}
\alias{kmg2_discretebar}
\alias{kmg2_discretebar-class}
\alias{plotDiscretebar,kmg2_discretebar-method}
\alias{kmg2_plot_discretebar}
\title{Bar chart for discrete variables}
\description{
  This function calls an R Commander plug-in window for Bar charts for discrete variables.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
