\name{kmg2_gettextRcmdr}
\alias{kmg2_gettextRcmdr}
\title{A gettextRcmdr wrapper function}
\description{
  This function is a gettextRcmdr wrapper function for this package.
}
\usage{
  kmg2_gettextRcmdr(...)
}
\arguments{
  \item{...}{arguments passed to gettext function}
}
\examples{
kmg2_gettextRcmdr("gettext")
}
\seealso{
  \code{\link[Rcmdr:Rcmdr.Utilities]{Rcmdr.Utilities}}
}
\keyword{hplot}
