\name{kmg2_hist-class}
\docType{class}
\alias{kmg2_hist}
\alias{kmg2_hist-class}
\alias{plotHist,kmg2_hist-method}
\alias{kmg2_plot_hist}
\title{Histogram}
\description{
  This function calls an R Commander plug-in window for Histograms.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}},
  \code{\link{kmg2_hist_df}}
}
\keyword{hplot}
