\name{kmg2_km-class}
\docType{class}
\alias{kmg2_km}
\alias{kmg2_km-class}
\alias{plotKM,kmg2_km-method}
\alias{kmg2_plot_km}
\title{Kaplan-Meier plot}
\description{
  This function calls an R Commander plug-in window for Kaplan-Meier plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}},
  \code{\link[survival:survfit]{survfit}},
  \code{\link{kmg2_km_df}}
}
\keyword{hplot}
