\name{kmg2_line-class}
\docType{class}
\alias{kmg2_line}
\alias{kmg2_line-class}
\alias{plotLine,kmg2_line-method}
\alias{kmg2_plot_line}
\title{Line chart}
\description{
  This function calls an R Commander plug-in window for line charts.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}},
  \code{\link{kmg2_line_df}}
}
\keyword{hplot}
