\name{kmg2_pie-class}
\docType{class}
\alias{kmg2_pie}
\alias{kmg2_pie-class}
\alias{plotPie,kmg2_pie-method}
\alias{kmg2_plot_pie}
\title{Pie chart}
\description{
  This function calls an R Commander plug-in window for pie charts.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
