\name{kmg2_qq-class}
\docType{class}
\alias{kmg2_qq}
\alias{kmg2_qq-class}
\alias{plotQQ,kmg2_qq-method}
\alias{kmg2_plot_qq}
\title{Q-Q plot}
\description{
  This function calls an R Commander plug-in window for Q-Q plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
