\name{kmg2_scatter-class}
\docType{class}
\alias{kmg2_scatter}
\alias{kmg2_scatter-class}
\alias{plotScatter,kmg2_scatter-method}
\alias{kmg2_plot_scatter}
\title{Scatter plot}
\description{
  This function calls an R Commander plug-in window for scatter plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
