\name{kmg2_scattermatrix-class}
\docType{class}
\alias{kmg2_scattermatrix}
\alias{kmg2_scattermatrix-class}
\alias{plotScattermatrix,kmg2_scattermatrix-method}
\alias{kmg2_plot_scattermatrix}
\title{Scatter plot matrix}
\description{
  This function calls an R Commander plug-in window for scatter plot matrix.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{hplot}
