\name{kmg2_theme}
\alias{theme_gray_natrisk}
\alias{theme_gray_natrisk21}
\alias{theme_bw_natrisk}
\alias{theme_bw_natrisk21}
\title{Extend graphics theme}
\description{
  These functions are modified graphics themes for ggplot2 package.
}
\usage{
  theme_gray_natrisk(base_size = 20, base_family = "")
}
\arguments{
  \item{base_size}{base font size}
  \item{base_family}{font family name}
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{color}
