\name{kmg2_variablesbox-class}
\docType{class}
\alias{kmg2_variablesbox}
\alias{kmg2_variablesbox-class}
\alias{front,kmg2_variablesbox-method}
\title{An R5 reference class of \code{variablesbox}}
\description{
  This class is a \code{variablesbox} class of this package.
}
\keyword{hplot}
