% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.r
\docType{package}
\name{RcmdrPlugin.KMggplot2-package}
\alias{KMggplot2}
\alias{RcmdrPlugin.KMggplot2}
\alias{RcmdrPlugin.KMggplot2-package}
\title{Rcmdr Plug-in for Data Visualization with 'ggplot2'}
\description{
A GUI front-end for 'ggplot2' supports Kaplan-Meier plot, histogram,
Q-Q plot, box plot, errorbar plot, scatter plot, line chart, pie chart,
bar chart, contour plot, and distribution plot.
}
\author{
Author: Triad sou. and Kengo NAGASHIMA
Maintainer: Triad sou. \email{triadsou@gmail.com}
}
\seealso{
\code{\link[Rcmdr:Rcmdr]{Rcmdr}},
\code{\link[ggplot2:ggplot]{ggplot2}},
\code{\link[survival:survfit]{survfit}},
\code{\link[RColorBrewer:RColorBrewer]{RColorBrewer}}
\code{\link[ggthemes:ggthemes]{ggthemes}}
}
\keyword{hplot}
\keyword{package}

