\name{windowDiscretebar}
\alias{windowDiscretebar}
\title{Wrapper Function of Discrete Bar Charts Subclass}
\usage{
  windowDiscretebar()
}
\description{
  \code{windowDiscretebar} function is a wrapper function
  of \code{discretebar} class for the R-commander menu bar.
}
\keyword{hplot}

