\name{windowDistHyper}
\alias{windowDistHyper}
\title{Wrapper Function of Hypergeometric Distribution Plot Subclass}
\usage{
  windowDistHyper()
}
\description{
  \code{windowDistHyper} function is a wrapper function of
  \code{dist} class for the R-commander menu bar.
}
\keyword{hplot}

