\name{windowDistNbinom}
\alias{windowDistNbinom}
\title{Wrapper Function of Negative Binomial Distribution Plot Subclass}
\usage{
  windowDistNbinom()
}
\description{
  \code{windowDistNbinom} function is a wrapper function of
  \code{dist} class for the R-commander menu bar.
}
\keyword{hplot}

