\name{windowDistWeibull}
\alias{windowDistWeibull}
\title{Wrapper Function of Weibull Distribution Plot Subclass}
\usage{
  windowDistWeibull()
}
\description{
  \code{windowDistWeibull} function is a wrapper function
  of \code{dist} class for the R-commander menu bar.
}
\keyword{hplot}

