\docType{methods}
\name{getXlab,plot_base-method}
\alias{getXlab,plot_base-method}
\title{Get \code{xlab}}
\arguments{
  \item{parms}{List of objects; graphics settings
  parameters.}
}
\value{
  \describe{ \item{xlab}{Character; a code snippet for the
  x-axis label.} }
}
\description{
  \code{getXlab} method gets the x-axis label.
}
\seealso{
  Other plot: \code{\link{box-class}},
  \code{\link{checkError,plot_base-method}},
  \code{\link{checkTheme,plot_base-method}},
  \code{\link{checkVariable,plot_base-method}},
  \code{\link{contour-class}},
  \code{\link{discretebar-class}},
  \code{\link{dist-class}}, \code{\link{factorize-class}},
  \code{\link{getCoord,plot_base-method}},
  \code{\link{getFacet,plot_base-method}},
  \code{\link{getGeom,plot_base-method}},
  \code{\link{getGgplot,plot_base-method}},
  \code{\link{getHelp,plot_base-method}},
  \code{\link{getMain,plot_base-method}},
  \code{\link{getMessage,plot_base-method}},
  \code{\link{getOpts,plot_base-method}},
  \code{\link{getParms,plot_base-method}},
  \code{\link{getPlot,plot_base-method}},
  \code{\link{getScale,plot_base-method}},
  \code{\link{getTheme,plot_base-method}},
  \code{\link{getWindowTitle,plot_base-method}},
  \code{\link{getYlab,plot_base-method}},
  \code{\link{getZlab,plot_base-method}},
  \code{\link{hist-class}}, \code{\link{km-class}},
  \code{\link{line-class}}, \code{\link{pie-class}},
  \code{\link{plot_base-class}},
  \code{\link{plotWindow,plot_base-method}},
  \code{\link{qq-class}},
  \code{\link{registRmlist,plot_base-method}},
  \code{\link{removeRmlist,plot_base-method}},
  \code{\link{savePlot,plot_base-method}},
  \code{\link{scatter-class}},
  \code{\link{scattermat-class}},
  \code{\link{setBack,plot_base-method}},
  \code{\link{setDataframe,plot_base-method}},
  \code{\link{setFront,plot_base-method}}
}
\keyword{hplot}

