% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.KMggplot2-package.r
\docType{package}
\name{RcmdrPlugin.KMggplot2-package}
\alias{RcmdrPlugin.KMggplot2-package}
\alias{RcmdrPlugin.KMggplot2}
\alias{KMggplot2}
\title{R Commander Plug-in for Data Visualization with 'ggplot2'}
\description{
A GUI front-end for 'ggplot2' supports Kaplan-Meier plot, histogram,
Q-Q plot, box plot, errorbar plot, scatter plot, line chart, pie chart,
bar chart, contour plot, and distribution plot.
}
\seealso{
\code{\link[Rcmdr:Rcmdr]{Rcmdr}},
\code{\link[ggplot2:ggplot]{ggplot2}},
\code{\link[survival:survfit]{survfit}},
\code{\link[RColorBrewer:RColorBrewer]{RColorBrewer}}
\code{\link[ggthemes:ggthemes]{ggthemes}}
\code{\link[scales:scales]{scales}}
}
\author{
\strong{Maintainer}: Kengo Nagashima \email{nshi1201@gmail.com} (\href{https://orcid.org/0000-0003-4529-9045}{ORCID})

}
\keyword{hplot}
\keyword{package}
