\name{RcmdrPlugin.KMggplot2}
\alias{RcmdrPlugin.KMggplot2}
\docType{package}
\title{
An R Commander plug-in for Kaplan-Meier plots and other plots by using the ggplot2 package
}
\description{
This package is an R Commander plug-in for Kaplan-Meier plots and other plots by using the ggplot2 package in GUI interface.
}
\author{
Author: Triad sou. and Kengo NAGASHIMA

Maintainer: Triad sou. \email{triadsou@gmail.com}
}
\keyword{hplot}
