\name{theme_bw_natrisk}
\alias{theme_bw_natrisk}
\alias{theme_bw_natrisk21}
\title{A Modified Black and White Theme for Kaplan-Meier Plots}
\usage{
  theme_bw_natrisk(base_size = 20, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{font family name}
}
\description{
  These functions are modified graphics themes for
  Kaplan-Meier plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{color}

