\name{theme_simple}
\alias{theme_simple}
\title{A Simple Black and White Theme}
\usage{
  theme_simple(base_size = 16, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{font family name}
}
\description{
  A Simple Black and White Theme
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{color}

