\name{theme_rect2}
\alias{theme_rect2}
\title{Theme element: modified rectangle (top and right borders removed).}
\usage{
  theme_rect2(colour = "black", size = 0.5, linetype = 1)
}
\arguments{
  \item{colour}{border color}

  \item{size}{border size}

  \item{linetype}{border linetype}
}
\description{
  Most often used for backgrounds and borders.
}
\seealso{
  \code{\link{polylineGrob}} for underlying grid function
}
\keyword{dplot}

