\name{gettextKmg2}
\alias{gettextKmg2}
\title{A \code{gettextRcmdr} Wrapper Function}
\usage{
  gettextKmg2(...)
}
\arguments{
  \item{...}{arguments passed to gettext function}
}
\description{
  This function is a \code{gettextRcmdr} wrapper function
  for this package.
}
\examples{
gettextKmg2("gettext")
}
\seealso{
  \code{\link[Rcmdr:Rcmdr.Utilities]{Rcmdr.Utilities}}
}
\keyword{documentation}

