\name{theme_simple}
\alias{theme_simple}
\alias{theme_simple_natrisk}
\alias{theme_simple_natrisk21}
\title{A Simple Black and White Themes}
\usage{
  theme_simple(base_size = 16, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{font family name}
}
\description{
  These functions are modified graphics themes for
  Kaplan-Meier plots.
}
\seealso{
  \code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{color}

